target: allow ALUA setup for some passthrough backends
This patch allows passthrough backends to use the core/base LIO
ALUA setup and state checks, but still handle the execution
of
commands.

This will allow the target_core_user module to
execute STPG and RTPG
in userspace, and not have to duplicate the ALUA
state checks, path
information (needed so we can check if
		command is executable on
		specific paths) and
setup (rtslib sets/updates the configfs ALUA
		interface like it does
		for iblock or file).

For STPG, the target_core_user userspace
daemon, tcmu-runner will
still execute the STPG, and to
update the core/base LIO state it
will use the existing configfs
interface. For RTPG, tcmu-runner
will loop over configfs
and/or cache the state.

Signed-off-by: Mike Christie <mchristi@redhat.com>
Signed-off-by: Nicholas Bellinger <nab@linux-iscsi.org>