tcmu: fix dev root_udev_waiter_mutex errors
This fixes the following bugs:

1. You cannot sleep after calling prepare_to_wait, so you
cannot use a mutex.

2. If the unmap thread is running and we have not done
the prepare_to_wait when a wake_up(&unmap_wait)
is done, then we will miss the wake up. We need to check
if a waiter has been added after setting the process state
and before calling schedule.

Note: there is a similar race with tcmu_irqconrol. One of the
next patches will fix that code path.

Signed-off-by: Mike Christie <mchristi@redhat.com>